chore: code cleanup and configuration options
